+++ CHANGELOG.md
@@ -3,6 +3,21 @@
 From 2019 onwards, all notable changes to tarpaulin will be documented in this
 file.
 
+## [0.36.0] 2026-06-29
+### Added
+- Adding support for custom cargo runners. This will be the basis of wasm and ebpf support
+- Basic feature gate evaluation to skip source analysis on ignored modules
+- Ptrace support for linux aarch64 (64 bit ARM)
+
+### Changed
+- Performance improvements in llvm profraw parsing
+
+### Fixed
+- Issue with config file and CLI merging that lead to CLI --engine being ignored if a config file is used
+
+### Fixed
+- Issue with config file and CLI merging that lead to CLI --engine being ignored if a config file is used
+
 ## [0.35.5] 2026-06-21
 ### Added
 - Cargo 1.94 onward the `CARGO_BIN_EXE_` env vars are now set at test runtime, replicate this when running under tarpaulin #1833